[POC] Offchain hardcoded story data oracle
A template offchain event listener that listens to the story protocol's smart contract event and write back data to the smart contract.
The code is for POC only. It's not intended for production use. To use it, replace the signer and the provider configuration with your own.

Tested in goerli testnet.

Issue #3